feat: Add integration description to replayed request ENV
Using request customizer add integration description at
`X_PACT_DESCRIPTION` to rack request env, similar to
`X_PACT_PROVIDER_STATES`. Remove this env in last middleware before rack
reversed proxy.

Additionally, support v3 provider states as a list with name and
parameters.
