+++ lib/pact/provider_verifier/custom_middleware.rb
@@ -1,5 +1,3 @@
-require 'ostruct'
-
 module Pact
   module ProviderVerifier
     class CustomMiddleware
@@ -23,7 +21,9 @@ module Pact
 
       def provider_states_from(env)
         if env["X_PACT_PROVIDER_STATES"]
-          env["X_PACT_PROVIDER_STATES"].collect{ | provider_state| OpenStruct.new(provider_state) }
+          env["X_PACT_PROVIDER_STATES"].collect do | provider_state|
+            Pact::ProviderState.from_hash(provider_state)
+          end
         else
           []
         end
